[packml_sm] Clearer dependency. Remove rqt.
Toward https://github.com/ros-industrial-consortium/packml/issues/23, I'm learning the code. If I understand correctly, packml_sm package seems to be an implementation of `packml` and the rest in packml repo (i.e. `packml_{gui, msgs, ros}`) are ROS and/or rqt binding of it.

If that is correct, it'd be cleaner and easier for maintenance for packml_sm to be ROS/rqt-agnostic.
Although with this PR `roscpp` remains in use, it looks to be relatively easy to remove roscpp as well. If that's possible packml_sm can be ROS-agnostic, more standalone library.